pyproject.toml
[tool.poetry]
name = "product-search-microservice"
version = "0.1.0"
description = "A gRPC microservice for searching products using Google Programmable Search Engine and Playwright"
authors = ["Arijit Roy arijit.roy@zageno.com"]
readme = "README.md"
packages = [{include = "productlookup", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
protobuf = "^4.25.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
playwright = "^1.41.1"
python-dotenv = "^1.0.1"
boto3 = "^1.34.0"
aiohttp = "^3.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "productlookup.main:serve"